(fix) Resolved merge erros, Added test folder with example files, Fixed git ignore not to add generated files, Fixed font size for Onboarding page
