+++ Mobile/starter_project_dartsmiths/.gitignore
@@ -42,3 +42,7 @@ app.*.map.json
 /android/app/debug
 /android/app/profile
 /android/app/release
+/macos/
+/windows/
+/linux/
+/ios/
