+++ Mobile/starter_project_dartsmiths/lib/features/onboarding/presentation/widgets/onboarding_stepper_box.dart
