+++ Mobile/starter_project_dartsmiths/pubspec.yaml
@@ -71,35 +71,21 @@ flutter:
   # "family" key with the font family name, and a "fonts" key with a
   # list giving the asset and other descriptors for the font. For
   # example:
-  fonts:
-    - family: Poppins
-      fonts:
-        - asset: assets/font/Poppins-Regular.ttf
-
-    - family: Poppins
-      fonts:
-        - asset: assets/font/Urbanist-VariableFont_wght.ttf
-          
-
-    # - family: Urbanist
-    #   fonts:
-    #     - asset: assets/fonts/Urbanist-Regular.ttf
         
-  #         style: italic
-  #   - family: Trajan Pro
-  #     fonts:
-  #       - asset: fonts/TrajanPro.ttf
-  #       - asset: fonts/TrajanPro_Bold.ttf
-  #         weight: 700
-  #
   fonts:
-    - family: Poppins
-      fonts:
-        - asset: assets/fonts/Poppins-Bold.ttf
-          weight: 600
-        - asset: assets/fonts/Poppins-Light.ttf
-          weight: 100
-        - asset: assets/fonts/Poppins-Medium.ttf
-          weight: 400
+  - family: Poppins
+    fonts:
+      - asset: assets/fonts/Poppins-Bold.ttf
+        weight: 600
+      - asset: assets/fonts/Poppins-Light.ttf
+        weight: 100
+      - asset: assets/fonts/Poppins-Medium.ttf
+        weight: 400
+      - asset: assets/fonts/Poppins-Regular.ttf
+
+  - family: Urbanist
+    fonts:
+      - asset: assets/fonts/Urbanist-VariableFont_wght.ttf
+        
   # For details regarding fonts from package dependencies,
   # see https://flutter.dev/custom-fonts/#from-packages
